- put shared functions into utils
