+++ mock_issues/src/pages/IssueInfo/IssueInfo.tsx
@@ -34,9 +34,7 @@ import {
 
 import MidHead from "../../components/MidHead/MidHead";
 import { FetchBaseQueryError } from "@reduxjs/toolkit/dist/query";
-
-type MyProps = {};
-type MyState = { editOnClick: boolean };
+import { countRestTime } from "../../utils/shareFunctions";
 
 export default function IssueInfo() {
 	const navigate = useNavigate();
@@ -177,47 +175,6 @@ export default function IssueInfo() {
 				},
 			});
 	}, [stateUpdateInfo]);
-
-	function countRestTime(timeString) {
-		const time = new Date(timeString);
-		const timeNow = Date.now();
-		const diffTime = timeNow - time.getTime();
-
-		const diffDays = Math.floor(diffTime / (24 * 3600 * 1000));
-		let hours, minutes, seconds;
-		if (diffDays <= 0) {
-			const leave1 = diffTime % (24 * 3600 * 1000);
-			hours = Math.floor(leave1 / (3600 * 1000));
-			if (hours <= 0) {
-				const leave2 = leave1 % (3600 * 1000);
-				minutes = Math.floor(leave2 / (60 * 1000));
-				if (minutes <= 0) {
-					const leave3 = leave2 % (60 * 1000);
-					seconds = Math.round(leave3 / 1000);
-					return `${seconds} seconds ago`;
-				} else {
-					return `${minutes} minutes ago`;
-				}
-			} else {
-				return `${hours} hours ago`;
-			}
-		} else if (diffDays <= 30) {
-			return `${diffDays} days ago`;
-		} else {
-			time.toLocaleString("default", { month: "short" });
-
-			time.toLocaleString("en-GB", {
-				day: "numeric",
-				month: "long",
-				year: "numeric",
-			});
-			return `on ${time.toLocaleString("en-GB", {
-				day: "numeric",
-				month: "long",
-				year: "numeric",
-			})}`;
-		}
-	}
 
 	return (
 		<div>
