+++ mock_issues/src/stories/Iconsstories/Label.tsx
@@ -1,31 +1,7 @@
 import styled from "styled-components";
+import { lightOrDark } from "utils/shareFunctions";
 
 export default function Label({ backgroundColor, labelName }) {
-	function lightOrDark(bgcolor) {
-		if (bgcolor.length === 7) {
-			const r = parseInt(bgcolor.slice(1, 3), 16);
-			const g = parseInt(bgcolor.slice(3, 5), 16);
-			const b = parseInt(bgcolor.slice(5, 7), 16);
-			const hsp = r * 0.3 + g * 0.6 + b * 0.1;
-			if (hsp > 127.5) {
-				return "#000000";
-			} else {
-				return "#ffffff";
-			}
-		} else if (bgcolor.length === 4) {
-			const r = parseInt(bgcolor.slice(1, 2) + bgcolor.slice(1, 2), 16);
-			const g = parseInt(bgcolor.slice(2, 3) + bgcolor.slice(2, 3), 16);
-			const b = parseInt(bgcolor.slice(3, 4) + bgcolor.slice(3, 4), 16);
-			const hsp = r * 0.3 + g * 0.6 + b * 0.1;
-			if (hsp > 127.5) {
-				return "#000000";
-			} else {
-				return "#ffffff";
-			}
-		}
-		if (bgcolor === "#ffff" || "#0000") return "#000000";
-	}
-
 	return (
 		<LabelA
 			labelColor={backgroundColor}
